vmm: move migration modules into folder
On-behalf-of: SAP philipp.schuster@sap.com
